+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # CHANGELOG
 
-## 3.3.1 (TBD)
+## 3.4.0 (TBD)
+* add support for purchase history records. New methods and properties:
+  - `BillingHelper.purchaseHistoryRecordsQueried`, `BillingHelper.initQueryPurchaseHistoryRecords` and `BillingHelper.getPurchaseHistoryRecords`
+  - `BillingEvent.QUERY_PURCHASES_HISTORY_RECORDS_COMPLETE` and `BillingEvent.QUERY_PURCHASES_HISTORY_RECORDS_FAILED`
+  - `queryPurchaseHistoryRecordsOnConnected` in `BillingHelper` constructor
 * fix `ProductDetails.getFormattedPriceDivided` to cover cases where price is formatted using comma and dot.
-* fix ´getPurchaseWithProductName´ now finds last known purchase matching a `productName`. 
-  This resolved issue where a subscription purchase gets updated but returned purchase wasn't a latest one.
+* rename `getPurchaseWithProductName` to `getPurchasesForProductName` as multiple purchases can exist for a single product name.
+  f.ex. when updating a subscription. Otherwise `.last()` is the one you're after.
 * bump Kotlin to 2.0.21, Gradle plugin to 8.7.3
 
 ## 3.3.0 (2024-11-11)